requirements.txt
requests>=2.28.2
Faker>=18.13.0
loguru>=0.7.0
web3>=6.5.0
python-dotenv>=1.0.0